Ports Unary to TRT dialect
- Ports the unary operations to the TRT dialect.

- Splits the Unary trace operations into separate classes instead of using
     a single class with an Enum.

- Splits the frontend ops corresponding to these into separate files to avoid
     very long files.
