+++ tripy/nvtripy/frontend/module/groupnorm.py
@@ -1,5 +1,5 @@
 #
-# SPDX-FileCopyrightText: Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -111,7 +111,7 @@ class GroupNorm(Module):
         """
         from nvtripy.frontend.ops.reduce import mean, var
         from nvtripy.frontend.ops.reshape import reshape
-        from nvtripy.frontend.ops.unary_elementwise import rsqrt
+        from nvtripy.frontend.ops.unary.rsqrt import rsqrt
 
         input_shape = x.shape
