+++ tripy/nvtripy/frontend/ops/unary/log.py
@@ -1,5 +1,4 @@
-#
-# SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-FileCopyrightText: Copyright (c) 2024 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,40 +12,33 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
-
 from nvtripy import export
+from nvtripy.frontend.ops import utils as op_utils
+from nvtripy.trace.ops.unary import Log
 from nvtripy.utils import wrappers
 
 
 @export.public_api(document_under="operations/functions")
 @wrappers.interface(
     dtype_constraints={"input": "T1", wrappers.RETURN_VALUE: "T1"},
-    dtype_variables={
-        "T1": ["float32", "float16", "bfloat16"],
-    },
+    dtype_variables={"T1": ["float32", "float16", "bfloat16"]},
 )
-def sigmoid(input: "nvtripy.Tensor") -> "nvtripy.Tensor":
-    r"""
-    Applies a logistic sigmoid function to each element of the input tensor:
-
-    :math:`\text{sigmoid}(x)_i = \frac{1}{1 + \exp{-x_i}}`
+def log(input: "nvtripy.Tensor") -> "nvtripy.Tensor":
+    """
+    Computes the elementwise natural logarithm (base e) of the elements of the input tensor.
 
     Args:
         input: The input tensor.
 
     Returns:
-        A tensor of the same shape as the input.
+        A new tensor of the same shape as the input tensor.
 
     .. code-block:: python
         :linenos:
 
-        input = tp.Tensor([1., 2., 3., 4.], dtype=tp.float32)
-        output = tp.sigmoid(input)
+        input = tp.arange(1, 3, dtype=tp.float32)
+        output = tp.log(input)
 
-        t = torch.tensor([1, 2, 3, 4], dtype=torch.float32) # doc: omit
-        assert tp.allclose(output, tp.Tensor(torch.nn.functional.sigmoid(t)))
+        assert tp.allclose(output, tp.Tensor(np.log(cp.from_dlpack(input).get())))
     """
-    from nvtripy.frontend.ops.unary.exp import exp
-
-    return 1.0 / (1.0 + exp(-1.0 * input))
+    return op_utils.create_op(Log, [input])
